@@ -1,5 +1,10 @@
 module github.com/Djarvur/go-err113
 
-go 1.13
+go 1.23.0
 
-require golang.org/x/tools v0.0.0-20200324003944-a576cf524670
+require golang.org/x/tools v0.36.0
+
+require (
+	golang.org/x/mod v0.27.0 // indirect
+	golang.org/x/sync v0.16.0 // indirect
+)
