@@ -3,8 +3,9 @@ package err113_test
 import (
 	"testing"
 
-	"github.com/Djarvur/go-err113"
 	"golang.org/x/tools/go/analysis/analysistest"
+
+	"github.com/Djarvur/go-err113"
 )
 
 func TestErr113(t *testing.T) {
